+++ src/main/java/com/example/fretesmartofi/repository/ICliente.java
@@ -3,5 +3,5 @@ package com.example.fretesmartofi.repository;
 import com.example.fretesmartofi.model.ClienteModel;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ICliente extends CrudRepository<ClienteModel, Integer> {
+public interface ICliente extends CrudRepository<ClienteModel, Long> {
 }
